update libft for further use
I change format of srcs files so that it would be easy to spotlight diff
and change fclean by dividing line so that make -j run will not remove
and compile simultaneously.

I update strtrim and itoa with better function name and atoi to produce
correct long max and min when input value is out of long range.

I update lst functions with exceptions mostly when **lst, *lst, *f, *del
is null pointer.